+++ perception/mrpython-pcl/src/mrpython_pcl/ros/LidarRosWrapper.py
@@ -7,13 +7,13 @@ import rospy
 import numpy as np
 import numpy.typing as npt
 from sensor_msgs.msg import PointCloud2
+from tf_helper.MarkerViz import MarkerViz  # pylint: disable=no-name-in-module, import-error
 
 from ..LidarPipeline.LidarPipeline import LidarPipeline
 from .Serializers import npToRos, npConesToRos, rosToPcl
-from .MarkerViz import MarkerViz
 
 
-class LidarRosWrapper(LidarPipeline):  # type: ignore[misc]
+class LidarRosWrapper(LidarPipeline):
     """
     A ros wrapper for the lidar pipeline class
     It allows the class to read sensor_msgs.msg.PointCloud2 point cloud
